tiny cleanup of comments and removed redundant cast the cast is needed in C# and it landed there the last time I flowed code backwards from the C# to the C to get comments.
I'm not actually typing the  comments 4 times :D

+++ valid.c
@@ -28,8 +28,8 @@ int valid(BOARD board, int is_white, int current_depth) {
       // row if you were to place a piece at column i.  So this says
       // if the effect of placing a piece at column i is only that the
       // piece at column i appears, then nothing flips, so it's not valid.
-      // remember me is in the high bits and him is in the low bits
-      // so we place onto the high bits.  And d1 has the current bit mask
+      // remember 'me' is in the high bits and 'him' is in the low bits
+      // So we place onto the high bits, hence mask << 8.
 
       if ((row | (mask << 8)) != flip_table[row][i]) {
         push_move(i, y, current_depth);
@@ -124,10 +124,8 @@ int valid(BOARD board, int is_white, int current_depth) {
     for (int i = y - 1; i >= 0; i--) {
       uint8_t h = him[i];
       uint8_t m = me[i];
-      uint32_t d0 =
-          ((uint8_t)(h >> (y - i)) << 16) | ((uint8_t)(h << (y - i)) << 8) | h;
-      uint32_t d1 =
-          ((uint8_t)(m >> (y - i)) << 16) | ((uint8_t)(m << (y - i)) << 8) | m;
+      uint32_t d0 = ((h >> (y - i)) << 16) | ((h << (y - i)) << 8) | h;
+      uint32_t d1 = ((m >> (y - i)) << 16) | ((m << (y - i)) << 8) | m;
 
       y0 = (~y1 & d0) | (y0 & (y1 | d1));
       y1 |= d1 | ~d0;
